Enhance conversation reply functionality and update project settings
- Added permissions for `conversation_reply` collection with actions for read, create, update, and delete.
- Updated GraphQL schema to include queries and mutations for `conversation_reply`.
- Introduced new fields in the project settings for enabling AI-powered replies and customizing reply prompts.
- Refactored frontend components to integrate new reply features, including UI elements for enabling replies and managing prompts.
- Improved backend logic for generating and storing conversation replies, ensuring proper handling of response content.
