+++ echo/directus/sync/snapshot/collections/conversation_reply.json
@@ -1,0 +1,28 @@
+{
+  "collection": "conversation_reply",
+  "meta": {
+    "accountability": "all",
+    "archive_app_filter": true,
+    "archive_field": null,
+    "archive_value": "archived",
+    "collapse": "open",
+    "collection": "conversation_reply",
+    "color": null,
+    "display_template": null,
+    "group": null,
+    "hidden": false,
+    "icon": null,
+    "item_duplication_fields": null,
+    "note": null,
+    "preview_url": null,
+    "singleton": false,
+    "sort": null,
+    "sort_field": "sort",
+    "translations": null,
+    "unarchive_value": "draft",
+    "versioning": false
+  },
+  "schema": {
+    "name": "conversation_reply"
+  }
+}
