+++ echo/directus/sync/snapshot/fields/conversation/processing_error.json
@@ -15,7 +15,7 @@
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 14,
+    "sort": 15,
     "special": null,
     "translations": null,
     "validation": null,
