+++ echo/directus/sync/snapshot/fields/conversation/processing_started_at.json
@@ -15,7 +15,7 @@
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 15,
+    "sort": 16,
     "special": null,
     "translations": null,
     "validation": null,
