+++ echo/directus/sync/snapshot/fields/conversation/processing_status.json
@@ -15,7 +15,7 @@
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 13,
+    "sort": 14,
     "special": null,
     "translations": null,
     "validation": null,
