+++ echo/directus/sync/snapshot/fields/conversation/project_chats.json
@@ -15,7 +15,7 @@
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 18,
+    "sort": 19,
     "special": [
       "m2m"
     ],
