+++ echo/directus/sync/snapshot/fields/conversation/summary.json
@@ -30,7 +30,7 @@
     },
     "readonly": false,
     "required": false,
-    "sort": 17,
+    "sort": 18,
     "special": null,
     "translations": null,
     "validation": null,
