+++ echo/directus/sync/snapshot/fields/conversation_reply/content_text.json
@@ -1,21 +1,21 @@
 {
-  "collection": "conversation",
-  "field": "processing_completed_at",
-  "type": "timestamp",
+  "collection": "conversation_reply",
+  "field": "content_text",
+  "type": "text",
   "meta": {
-    "collection": "conversation",
+    "collection": "conversation_reply",
     "conditions": null,
     "display": null,
     "display_options": null,
-    "field": "processing_completed_at",
+    "field": "content_text",
     "group": null,
-    "hidden": true,
-    "interface": null,
+    "hidden": false,
+    "interface": "input-multiline",
     "note": null,
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 17,
+    "sort": 4,
     "special": null,
     "translations": null,
     "validation": null,
@@ -23,9 +23,9 @@
     "width": "full"
   },
   "schema": {
-    "name": "processing_completed_at",
-    "table": "conversation",
-    "data_type": "timestamp with time zone",
+    "name": "content_text",
+    "table": "conversation_reply",
+    "data_type": "text",
     "default_value": null,
     "max_length": null,
     "numeric_precision": null,
