+++ echo/directus/sync/snapshot/fields/conversation_reply/date_created.json
@@ -1,30 +1,32 @@
 {
-  "collection": "conversation",
-  "field": "processing_completed_at",
+  "collection": "conversation_reply",
+  "field": "date_created",
   "type": "timestamp",
   "meta": {
-    "collection": "conversation",
+    "collection": "conversation_reply",
     "conditions": null,
     "display": null,
     "display_options": null,
-    "field": "processing_completed_at",
+    "field": "date_created",
     "group": null,
-    "hidden": true,
+    "hidden": false,
     "interface": null,
     "note": null,
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 17,
-    "special": null,
+    "sort": 8,
+    "special": [
+      "date-created"
+    ],
     "translations": null,
     "validation": null,
     "validation_message": null,
     "width": "full"
   },
   "schema": {
-    "name": "processing_completed_at",
-    "table": "conversation",
+    "name": "date_created",
+    "table": "conversation_reply",
     "data_type": "timestamp with time zone",
     "default_value": null,
     "max_length": null,
