+++ echo/directus/sync/snapshot/fields/conversation_reply/id.json
@@ -1,39 +1,41 @@
 {
-  "collection": "conversation",
-  "field": "processing_completed_at",
-  "type": "timestamp",
+  "collection": "conversation_reply",
+  "field": "id",
+  "type": "uuid",
   "meta": {
-    "collection": "conversation",
+    "collection": "conversation_reply",
     "conditions": null,
     "display": null,
     "display_options": null,
-    "field": "processing_completed_at",
+    "field": "id",
     "group": null,
     "hidden": true,
-    "interface": null,
+    "interface": "input",
     "note": null,
     "options": null,
-    "readonly": false,
+    "readonly": true,
     "required": false,
-    "sort": 17,
-    "special": null,
+    "sort": 1,
+    "special": [
+      "uuid"
+    ],
     "translations": null,
     "validation": null,
     "validation_message": null,
     "width": "full"
   },
   "schema": {
-    "name": "processing_completed_at",
-    "table": "conversation",
-    "data_type": "timestamp with time zone",
+    "name": "id",
+    "table": "conversation_reply",
+    "data_type": "uuid",
     "default_value": null,
     "max_length": null,
     "numeric_precision": null,
     "numeric_scale": null,
-    "is_nullable": true,
-    "is_unique": false,
+    "is_nullable": false,
+    "is_unique": true,
     "is_indexed": false,
-    "is_primary_key": false,
+    "is_primary_key": true,
     "is_generated": false,
     "generation_expression": null,
     "has_auto_increment": false,
