+++ echo/directus/sync/snapshot/fields/conversation_reply/type.json
@@ -1,21 +1,21 @@
 {
-  "collection": "conversation",
-  "field": "processing_completed_at",
-  "type": "timestamp",
+  "collection": "conversation_reply",
+  "field": "type",
+  "type": "string",
   "meta": {
-    "collection": "conversation",
+    "collection": "conversation_reply",
     "conditions": null,
     "display": null,
     "display_options": null,
-    "field": "processing_completed_at",
+    "field": "type",
     "group": null,
-    "hidden": true,
-    "interface": null,
+    "hidden": false,
+    "interface": "input",
     "note": null,
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 17,
+    "sort": 5,
     "special": null,
     "translations": null,
     "validation": null,
@@ -23,11 +23,11 @@
     "width": "full"
   },
   "schema": {
-    "name": "processing_completed_at",
-    "table": "conversation",
-    "data_type": "timestamp with time zone",
+    "name": "type",
+    "table": "conversation_reply",
+    "data_type": "character varying",
     "default_value": null,
-    "max_length": null,
+    "max_length": 255,
     "numeric_precision": null,
     "numeric_scale": null,
     "is_nullable": true,
