+++ echo/directus/sync/snapshot/fields/project/get_reply_prompt.json
@@ -1,21 +1,21 @@
 {
-  "collection": "conversation",
-  "field": "processing_completed_at",
-  "type": "timestamp",
+  "collection": "project",
+  "field": "get_reply_prompt",
+  "type": "text",
   "meta": {
-    "collection": "conversation",
+    "collection": "project",
     "conditions": null,
     "display": null,
     "display_options": null,
-    "field": "processing_completed_at",
+    "field": "get_reply_prompt",
     "group": null,
-    "hidden": true,
-    "interface": null,
+    "hidden": false,
+    "interface": "input-multiline",
     "note": null,
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 17,
+    "sort": 24,
     "special": null,
     "translations": null,
     "validation": null,
@@ -23,9 +23,9 @@
     "width": "full"
   },
   "schema": {
-    "name": "processing_completed_at",
-    "table": "conversation",
-    "data_type": "timestamp with time zone",
+    "name": "get_reply_prompt",
+    "table": "project",
+    "data_type": "text",
     "default_value": null,
     "max_length": null,
     "numeric_precision": null,
