+++ echo/directus/sync/snapshot/fields/project/is_get_reply_enabled.json
@@ -1,32 +1,34 @@
 {
-  "collection": "conversation",
-  "field": "processing_completed_at",
-  "type": "timestamp",
+  "collection": "project",
+  "field": "is_get_reply_enabled",
+  "type": "boolean",
   "meta": {
-    "collection": "conversation",
+    "collection": "project",
     "conditions": null,
     "display": null,
     "display_options": null,
-    "field": "processing_completed_at",
+    "field": "is_get_reply_enabled",
     "group": null,
-    "hidden": true,
-    "interface": null,
+    "hidden": false,
+    "interface": "boolean",
     "note": null,
     "options": null,
     "readonly": false,
     "required": false,
-    "sort": 17,
-    "special": null,
+    "sort": 23,
+    "special": [
+      "cast-boolean"
+    ],
     "translations": null,
     "validation": null,
     "validation_message": null,
     "width": "full"
   },
   "schema": {
-    "name": "processing_completed_at",
-    "table": "conversation",
-    "data_type": "timestamp with time zone",
-    "default_value": null,
+    "name": "is_get_reply_enabled",
+    "table": "project",
+    "data_type": "boolean",
+    "default_value": false,
     "max_length": null,
     "numeric_precision": null,
     "numeric_scale": null,
