+++ echo/frontend/src/lib/utils.ts
@@ -8,3 +8,28 @@ export function cn(...inputs: ClassValue[]) {
 export function capitalize(str: string) {
   return str.charAt(0).toUpperCase() + str.slice(1);
 }
+
+export const scrollToBottom = (
+  container: React.RefObject<HTMLDivElement | null>,
+) => {
+  container.current?.scrollIntoView({ behavior: "smooth" });
+};
+
+export const checkPermissionError = async () => {
+  try {
+    // @ts-expect-error microphone is not available?
+    const result = await navigator.permissions.query({ name: "microphone" });
+    if (result.state === "denied") {
+      return "denied" as const;
+    } else if (result.state === "prompt") {
+      return "prompt" as const;
+    } else if (result.state === "granted") {
+      return "granted" as const;
+    } else {
+      return "error" as const;
+    }
+  } catch (error) {
+    console.error("Error checking microphone permissions", error);
+    return "error" as const;
+  }
+};
