+++ echo/server/dembrane/reply_utils.py
@@ -1,5 +1,5 @@
-from logging import getLogger
 from typing import AsyncGenerator
+from logging import getLogger
 
 import litellm
 from pydantic import BaseModel
@@ -71,7 +71,9 @@ def build_conversation_transcript(conversation: dict) -> str:
     return transcript
 
 
-async def generate_reply_for_conversation(conversation_id: str, language: str) -> AsyncGenerator[str, None]:
+async def generate_reply_for_conversation(
+    conversation_id: str, language: str
+) -> AsyncGenerator[str, None]:
     conversation = directus.get_items(
         "conversation",
         {
@@ -232,37 +234,105 @@ async def generate_reply_for_conversation(conversation_id: str, language: str) -
 
     # Store the complete response
     accumulated_response = ""
-    
+
+    # Buffer to detect <response> tag which may be split across chunks
+    tag_buffer = ""
+    found_response_tag = False
+    # Also track closing tag to properly handle content
+    in_response_section = False
+
     # Stream the response
     response = await litellm.acompletion(
         model="anthropic/claude-3-5-sonnet-20240620",
-        messages=[{"role": "user", "content": prompt}, {"role": "assistant", "content": "<response>"},],
-        stream=True
+        messages=[
+            {"role": "user", "content": prompt},
+            {"role": "assistant", "content": ""},
+        ],
+        stream=True,
     )
+
+    # List of possible partial closing tags
+    partial_closing_patterns = [
+        "</r",
+        "</re",
+        "</res",
+        "</resp",
+        "</respo",
+        "</respon",
+        "</respons",
+        "</response",
+        "</response>",
+    ]
 
     async for chunk in response:
         if chunk.choices[0].delta.content:
             content = chunk.choices[0].delta.content
             accumulated_response += content
-            
-            # Remove any response tags that might appear
-            cleaned_content = content.replace("<response>", "").replace("</response>", "")
-            if cleaned_content.strip():  # Only yield non-empty content
-                yield cleaned_content
 
-    # After streaming is complete, store in Directus
+            if in_response_section:
+                # While inside the response section, check for closing tag
+                if any(partial in (tag_buffer + content) for partial in partial_closing_patterns):
+                    combined = tag_buffer + content
+
+                    for partial in partial_closing_patterns:
+                        partial_index = combined.find(partial)
+                        if partial_index != -1:
+                            to_yield = combined[:partial_index]
+                            if to_yield.strip():
+                                yield to_yield
+                            break  # Stop checking further once the first match is found
+
+                    # Stop streaming
+                    in_response_section = False
+                    tag_buffer = ""
+                else:
+                    # No closing tag found, continue yielding content
+                    if (tag_buffer + content).strip():
+                        yield tag_buffer + content
+                    tag_buffer = ""
+            else:
+                if not found_response_tag:
+                    # Append to buffer to handle split tags
+                    tag_buffer += content
+
+                    # Check if buffer contains <response>
+                    if "<response>" in tag_buffer:
+                        found_response_tag = True
+                        in_response_section = True
+
+                        # Extract content after the opening tag
+                        start_idx = tag_buffer.find("<response>") + len("<response>")
+                        content_after_tag = tag_buffer[start_idx:]
+
+                        # Reset buffer and yield content if any
+                        tag_buffer = ""
+                        if content_after_tag.strip():
+                            yield content_after_tag
+
+                    # If buffer gets too large without finding tag, trim it, keep only the last 20 characters
+                    if len(tag_buffer) > 100:
+                        tag_buffer = tag_buffer[-20:]
+
     try:
+        response_content = ""
+        if "<response>" in accumulated_response and "</response>" in accumulated_response:
+            start_idx = accumulated_response.find("<response>") + len("<response>")
+            end_idx = accumulated_response.find("</response>")
+            if start_idx < end_idx:
+                response_content = accumulated_response[start_idx:end_idx].strip()
+        else:
+            response_content = accumulated_response.strip()
+
         directus.create_item(
             "conversation_reply",
             item_data={
                 "conversation_id": current_conversation.id,
-                "content_text": accumulated_response.strip(),
+                "content_text": response_content,
                 "type": "assistant_reply",
             },
         )
     except Exception as e:
         logger.error(f"Failed to store reply in Directus: {e}")
-        # Continue since we've already streamed the response to the user
 
 
 if __name__ == "__main__":
